+++ models/content.js
@@ -1,0 +1,14 @@
+const mongoose = require("mongoose");
+
+const contentSchema = new mongoose.Schema({
+  title: String,
+  detail: String,
+  image: {
+    url: String,
+    filename: String,
+  },
+});
+
+const Content = mongoose.model("Content", contentSchema);
+
+module.exports = Content;
